requirements.txt
streamlit
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
selenium
beautifulsoup4
lxml
html5lib
requests
openai
groq
python-dotenv